+++ run.jl
@@ -1,0 +1,5 @@
+using Pkg
+Pkg.activate(@__DIR__)
+using Orchestrator
+
+Orchestrator.main()
